backend/requirements.txt
djitellopy==2.5.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
opencv-python>=4.10.0.84
numpy>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0
colorama>=0.4.6
pillow>=10.0.0
pydantic>=2.0.0
xai-sdk>=0.1.0
